requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astunparse==1.6.3
attrs==24.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.3.4
cryptography==42.0.5
cssselect==1.2.0
Django==5.0
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
django-tenant-schemas==1.12.0
django-tenants==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
filelock==3.13.4
flatbuffers==24.3.25
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.11.0
huggingface-hub==0.22.2
idna==3.7
inflection==0.5.1
joblib==1.4.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
libclang==18.1.1
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
ordered-set==4.1.0
packaging==24.0
pdfminer.six==20231228
phonenumbers==8.13.29
protobuf==4.25.3
psycopg2-binary==2.9.9
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyquery==2.0.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
roman==4.1
rpds-py==0.19.1
safetensors==0.4.3
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
termcolor==2.4.0
tokenizers==0.19.1
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
